Inherit Debian kernel configuration
I noticed that quite a few kernel configuration options which are
enabled by default in Debian (such as user namespaces, apparmor, etc)
are not enabled in the bubba3 kernel.  The simplest and most maintainable
solution I could find was to include the debian configuration files
in this repository and to merge those with the bubba3 configuration at
build time.

I have moved the bubba3 configuration from a patch to
debian/config/config.bubba3.  I have not changed this configuration at
all.  The debian configurations were copied from the debian source
package linux-4.19.37-5.

The config merge script already included in the official kernel
repository is a bit "noisy" (it prints warnings whenever a previous
value has been overridden), but I guess it is better to be a bit noisy
than not noticing that some comments ("# CONFIG_... is not set")
actually overrides values.